gp_bank_2/server/requirements.txt
tensorflow==2.12.0
python-dotenv==1.0.0
absl-py==1.4.0
aiohttp==3.8.4
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
boto3==1.28.32
botocore==1.31.32
bs4==0.0.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
DAWG-Python==0.7.2
docopt==0.6.2
exceptiongroup==1.1.3
fastapi==0.100.0
flatbuffers==23.5.26
frozenlist==1.4.0
gast==0.4.0
gensim==4.3.1
google-auth==2.22.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.57.0
h11==0.14.0
h5py==3.9.0
idna==3.4
importlib-metadata==6.8.0
inexactsearch==1.0.2
jmespath==1.0.1
joblib==1.3.2
keras==2.12.0
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
multidict==6.0.4
numpy==1.23.2
oauthlib==3.2.2
openai==0.27.9
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
protobuf==4.24.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.0.3
pydantic_core==2.3.0
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
PySocks==1.7.1
pyspellchecker==0.7.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
s3transfer==0.6.2
scikit-learn==1.3.0
scipy==1.11.1
silpa-common==0.3
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.4.1
spellchecker==0.4
starlette==0.27.0
tensorboard==2.12.0
termcolor==2.3.0
threadpoolctl==3.2.0
tqdm==4.65.0
trio==0.22.1
trio-websocket==0.10.3
typing==3.7.4.3
typing_extensions==4.6.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.1
Werkzeug==2.3.7
wsproto==1.2.0
yarl==1.9.2
zipp==3.16.2
